# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    A universal room reservation system designed for use in various institutions and organizations.

    Possible applications:
    - universities, schools – lecture halls, laboratories
    - companies, hotels – conference rooms
    - libraries, cultural centers – workshop rooms, event spaces
    - gyms and fitness clubs – exercise rooms
  title: RoomReservation
  version: 0.0.1
servers:
  - url: /api
...
paths:
...
  /tags:
    get:
      description: Retrieve a list of all available tags.
      operationId: listTags
      responses:
        "200":
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Tag'
          description: List of all tags
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected server error.
      summary: List all tags
      tags:
        - tags
      x-accepts:
        - application/json
      x-tags:
        - tag: tags
    post:
      description: Create a new tag.
      operationId: createTag
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createTag_request'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
          description: Tag created successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid input.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected server error.
      summary: Create a new tag
      tags:
        - tags
      x-content-type: application/json
      x-accepts:
        - application/json
      x-tags:
        - tag: tags
  /tags/{tag_id}:
    get:
      description: Retrieve details of a specific tag.
      operationId: getTag
      parameters:
        - description: ID of the tag to retrieve
          explode: false
          in: path
          name: tag_id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
          description: Tag details
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Requested resource was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected server error.
      summary: Get tag details
      tags:
        - tags
      x-accepts:
        - application/json
      x-tags:
        - tag: tags
    put:
      description: Update an existing tag.
      operationId: updateTag
      parameters:
        - description: ID of the tag to update
          explode: false
          in: path
          name: tag_id
          required: true
          schema:
            type: integer
          style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/updateTag_request'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Tag'
          description: Tag updated successfully
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid input.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Requested resource was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected server error.
      summary: Update a tag
      tags:
        - tags
      x-content-type: application/json
      x-accepts:
        - application/json
      x-tags:
        - tag: tags
    delete:
      description: Delete a tag.
      operationId: deleteTag
      parameters:
        - description: ID of the tag to delete
          explode: false
          in: path
          name: tag_id
          required: true
          schema:
            type: integer
          style: simple
      responses:
        "204":
          description: Tag deleted successfully
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Requested resource was not found.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unexpected server error.
      summary: Delete a tag
      tags:
        - tags
      x-accepts:
        - application/json
      x-tags:
        - tag: tags
...
components:
...
  schemas:
    AuditFields:
      description: Audit fields containing timestamps of creation and last update.
      properties:
        createdAt:
          description: Timestamp when the record was created (ISO 8601 format).
          example: 2025-05-24T12:34:56Z
          format: date-time
          type: string
        updatedAt:
          description: Timestamp of the last modification of the record.
          example: 2025-05-25T08:00:00Z
          format: date-time
          type: string
      type: object
...
    Tag:
      allOf:
        - $ref: '#/components/schemas/AuditFields'
        - description: A tag representing a feature or characteristic of a room.
          properties:
            id:
              description: Unique identifier of the tag.
              example: 1
              type: integer
            name:
              description: Name of the tag.
              example: projector
              type: string
            description:
              description: Optional description of what the tag represents.
              example: Room has a projector available
              type: string
          required:
            - name
          type: object
      example:
        id: 1
        name: projector
        description: Room has a projector available
        createdAt: 2025-05-24T12:34:56Z
        updatedAt: 2025-05-24T12:34:56Z
      type: object
    createTag_request:
      properties:
        name:
          description: Name of the tag
          example: whiteboard
          type: string
        description:
          description: Optional description of what the tag represents
          example: Room has a whiteboard
          type: string
      required:
        - name
      type: object
    updateTag_request:
      properties:
        name:
          description: Updated name of the tag
          example: smart-whiteboard
          type: string
        description:
          description: Updated description of the tag
          example: Room has an interactive smart whiteboard
          type: string
      type: object
...
    Error:
      description: Error response format.
      example:
        code: BadRequest
        message: Invalid input provided.
      properties:
        code:
          description: Error code identifier.
          example: BadRequest
          type: string
        message:
          description: Detailed error message.
          example: Invalid input provided.
          type: string
      type: object
...